Adding ESP32-S3 support (IDF v4.4.1)
This splits the io_mux_reg conversion up into board-specific files that
are included when `-d:esp32s3` or not.

The default case is the ESP32, to not break existing projects that do
not have a define yet.

This ports all of the extra GPIO pins, and re-enables some missing pins
for the S3 in the existing driver. This has been tested at least
somewhat on real hardware, and appears to be working well